Add api controller and routes

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -107,9 +107,9 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        /*if ($project->image) {
+        if ($project->image) {
             Storage::delete($project->image);
-        }*/
+        }
 
         $project->technologies()->detach();
